% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{\%at\%}
\alias{\%at\%}
\title{Evaluate expression x in envAns environment}
\usage{
x \%at\% envAns
}
\arguments{
\item{x}{expressions contained by {}}

\item{envAns}{an environment}
}
\value{

}
\description{
Evaluate expression x in envAns environment
}
\examples{
none
}
