% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{windowsService}
\alias{windowsService}
\title{Create windowService to fix different Windows problems}
\usage{
windowsService()
}
\value{

}
\description{
Create windowService to fix different Windows problems
}
\examples{
ws <- windowService()
ws$set_RconsoleLanguage2English()
}
