% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{write_utf8}
\alias{write_utf8}
\title{Safe write method based on https://kevinushey.github.io/blog/2018/02/21/string-encoding-and-r/}
\usage{
write_utf8(text, f = tempfile())
}
\arguments{
\item{text}{A charactor.}

\item{f}{A charactor.}
}
\value{

}
\description{
Safe write method based on https://kevinushey.github.io/blog/2018/02/21/string-encoding-and-r/
}
\examples{
none
}
